act: Initial commit
Create a 2D expansion vector with the cells representing the index at which they were expanded. 

A slight modification is added to the BFS Udacity code algorithm, which allows for lesser steps and identifying the goal directly when the goal cell is seen rather than when the goal cell is picked.